Consolidate list_eval_logs_async and add trio recover regression test
Pull in the remaining improvements from upstream PR #4625's final
revision on top of #4676:

- Promote the view server's genuinely-async lister into log/_file.py as
  _list_eval_logs_async: the public list_eval_logs_async() now lists
  natively async on remote filesystems (S3 fast path via
  AsyncFilesystem.iter_files with missing-bucket handling, async fsspec
  walk with Azure auth-error suppression) instead of blocking the event
  loop on sync fs.exists/fs.ls. async_filesystem and the _walk_* helpers
  move along with it, with lazy s3fs/botocore/_view.azure imports so
  they stay out of inspect_ai.log's import graph. #4676's concurrent
  filter (tg_collect + semaphore) is kept on top of the new core.
- _view/common.py now consumes the public lister, eliminating the two
  same-named implementations and their reciprocal "keep the two
  aligned" comments.
- Add test_recover_async_paths_trio: regular-CI trio coverage (via
  anyio.run(backend="trio")) of recover_eval_log_async() and
  _recoverable_eval_logs_async(), the path issue #171 started from.
- Add edge tests from #4625: async-matches-sync, filter-excludes-all,
  and missing-dir listings.

Fixes meridianlabs-ai/inspect_ai#179
